Fixing layout issues.
Adding call to the viewWillLayoutSubviews method to update the view
frames. This is needed if for example a full screen view controller is
presented and the device is rotated (only the top most view controller
get the rotation callbacks). So this will allow the views (master +
divider + detail) to be properly laid out always.